==> 11_Remove_Outliers.R <==
##%######################################################%##
#                                                          #
####        11. Additional test: Remove outliers        ####
#                                                          #
##%######################################################%##


# This script checks for outliers in the models, removes these from the 
# dataset and reruns the models to see if they have an effect on the results.


rm(list = ls())

# directories
moddir <- "5_RunLUIClimateModels/Including outliers/"
outdir <- "10_Additional_Tests/Outliers/"
if(!dir.exists(outdir)) dir.create(outdir)

# libraries
library(performance)
source("0_Functions.R")
library(influence.ME)
library(StatisticalModels)
library(cowplot)
library(sjPlot)
library(ggplot2)
library(predictsFunctions)
library(dplyr)

# load in models
load(paste0(moddir, "MeanAnomalyModelAbund_incoutliers.rdata")) # MeanAnomalyModelAbund
load(paste0(moddir, "MeanAnomalyModelRich_incoutliers.rdata"))  # MeanAnomalyModelRich


############################################################
#                                                          #
####               1. Assess outliers                   ####
#                                                          #
############################################################


# 1. abundance mean anomaly

# extract the data used to run the model
modelData <- MeanAnomalyModelAbund$data

# number of unique studies
length(unique(modelData$SS)) # 235

# rerun the model using lme4 to avoid errors when using the influence function
mod1 <- lmer(formula = LogAbund ~ LUI * StdTmeanAnomalyRS * Order + (1 | SS) + (1 | SSB), modelData)

# use influence function to get outliers and plot
alt.est.a <- influence(mod1, "SS") # takes a long time to run
plot(alt.est.a, which = "cook", sort = T) 
result1 <- cooks.distance(alt.est.a, sort = T)

result1 <- as.data.frame(result1)
result1$study <- rownames(result1)
View(result1[result1$V1 >= 1,])

write.csv(result1[result1$V1 >= 1,], file = paste0(outdir, "Abun_Cooks_distances.csv"), row.names = F)

abun_ss <- result1[result1$V1 >= 1, "study"]

# 3 studies are classed as outliers (cook's distance greater than 1)
# "AD1_2008__Billeter 6"    "HW1_2011__Summerville 1" "SC1_2005__Richardson 1" 

# Take a look at the outliers
test <- modelData[modelData$SS %in% abun_ss, ]
length(unique(test$SSBS)) # 203

table(test$Order)
# Coleoptera     Diptera   Hemiptera Hymenoptera Lepidoptera 
# 139                123          59         123         139 

# percentage of sites that are included in the outlying studies
table(test$Order)/table(modelData$Order) *100
# Coleoptera     Diptera   Hemiptera Hymenoptera Lepidoptera 
# 6.056645   17.571429    6.933020    3.938521    9.272849 

# assess land use of outlying studies
table(test$Order, test$LUI)
#                 Primary vegetation Secondary vegetation Agriculture_Low Agriculture_High
# Coleoptera                       0                  139               0                0
# Diptera                          0                   59              32               32
# Hemiptera                        0                   59               0                0
# Hymenoptera                      0                   59              32               32
# Lepidoptera                      0                  139               0                0



# 2. Richness mean anomaly

modelData <- MeanAnomalyModelRich$data

length(unique(modelData$SS)) # 253

# rerun the model using lme4 to avoid errors when using the influence function
mod2 <- glmer(formula = Species_richness~LUI * StdTmeanAnomalyRS * Order+(1|SS)+(1|SSB)+(1|SSBS), modelData, family = "poisson")

# assess outliers
alt.est.a <- influence(mod2, "SS") # takes a really long time to run
plot(alt.est.a, which = "cook", sort = T)
result2 <- as.data.frame(cooks.distance(alt.est.a, sort = T))
result2$SS <- rownames(result2)
View(result2[result2$V1 >= 1,])

write.csv(result2[result2$V1 >= 1,], file = paste0(outdir, "Rich_Cooks_distances.csv"), row.names = F)


# study IDs for those with cook's distance greater than 1
rich_ss <- result2[result2$V1 >= 1, "study"]

length(rich_ss) # 14

test <- modelData[modelData$SS %in% rich_ss, ] # 2207 rows
length(unique(test$SSBS)) # 803

hist(test$Species_richness)
range(test$Species_richness) # 0 - 161
quantile(test$Species_richness) #   0%  25%  50%  75% 100% 
                                   # 0    1    4   11  161 

# bring together information on outliers
rich_bal <- test %>% 
  group_by(SS) %>% 
  summarise(minrich = min(Species_richness),
            maxrich = max(Species_richness), 
            nLUI = length(unique(LUI)), 
            nOrders = length(unique(Order)),
            nSites = length(unique(SSBS)), 
            nrows = length(SSBS)) 

rich_bal <- inner_join(rich_bal, result2, by = "SS") %>% 
    arrange(desc(V1))

write.csv(rich_bal, paste0(outdir, "Outliers_data_summary.csv"), row.names = F)

table(droplevels(modelData[modelData$SS %in% rich_ss, "SS"]))

# number of site/order combinations, highest cooks distance at the top
# HW1_2011__Summerville 1 ......160
# AR1_2008__Basset 1............45
# SC1_2011__Meijer 1............456
# SC1_2005__Richardson 1........295
# SH1_2011__Todd 1 .............100
# AD1_2008__Billeter 6..........128
# TN1_2007__Bouyer 2 ...........192
# DI1_2013__Rousseau 2..........360  # here and above have cooks distance higher than 2 (total 1736)
# HP1_2006__Lachat 1............72
# SE1_2012__Poveda 2............85
# LH1_2008__Littlewood 1........152
# SE1_2012__Poveda 1............85
# AD1_2008__Franzen.............32
# SC1_2010__Marsh 1.............45

table(droplevels(modelData[modelData$SS %in% rich_ss, "SS"]), droplevels(modelData[modelData$SS %in% rich_ss, "Order"]))
#                         Coleoptera Diptera Hemiptera Hymenoptera Lepidoptera
# AR1_2008__Basset 1               9       9         9           9           9
# DI1_2013__Rousseau 2            72      72        72          72          72
# HP1_2006__Lachat 1              36       0        36           0           0
# HW1_2011__Summerville 1         80       0         0           0          80
# SC1_2005__Richardson 1          59      59        59          59          59
# SC1_2011__Meijer 1             152       0       152           0         152
# SE1_2012__Poveda 1              17      17        17          17          17
# SE1_2012__Poveda 2              17      17        17          17          17
# SH1_2011__Todd 1                20      20        20          20          20
# TN1_2007__Bouyer 2              64       0         0          64          64
# AD1_2008__Billeter 6             0      64         0          64           0
# AD1_2008__Franzen 1              0       0         0          16          16
# LH1_2008__Littlewood 1           0       0         0           0         152
# SC1_2010__Marsh 1                0       0         0           0          45
#                                                                                       
                       
table(droplevels(modelData[modelData$SS %in% rich_ss, "SS"]), droplevels(modelData[modelData$SS %in% rich_ss, "LUI"]))
#                         Primary vegetation Secondary vegetation Agriculture_Low Agriculture_High
# AR1_2008__Basset 1                       0                   45               0                0
# DI1_2013__Rousseau 2                     0                   90             180               90
# HP1_2006__Lachat 1                      24                   24               0               24
# HW1_2011__Summerville 1                  0                  160               0                0
# SC1_2005__Richardson 1                   0                  295               0                0
# SC1_2011__Meijer 1                     198                   81             177                0
# SE1_2012__Poveda 1                       0                    0               0               85
# SE1_2012__Poveda 2                       0                    0               0               85
# SH1_2011__Todd 1                         0                    0               0              100
# TN1_2007__Bouyer 2                      69                   39              48               36
# AD1_2008__Billeter 6                     0                    0              64               64
# AD1_2008__Franzen 1                      0                    4              28                0
# LH1_2008__Littlewood 1                   0                   38             114                0
# SC1_2010__Marsh 1                       15                   15               0               15
       
nrow(test) # 2207
table(test$Order)
# Coleoptera     Diptera   Hemiptera Hymenoptera Lepidoptera 
#        526         258         382         338         703 

table(test$Order)/table(modelData$Order) *100
# Coleoptera     Diptera   Hemiptera Hymenoptera Lepidoptera 
#   22.72138    35.83333    43.85763    10.23312    42.63190 

#test <- modelData[modelData$SS %in% rich_ss, ]

table(test$Order, test$LUI)
# 
#             Primary vegetation Secondary vegetation Agriculture_Low Agriculture_High
# Coleoptera                 101                  218             111               96
# Diptera                      0                   86              68              104
# Hemiptera                   78                  125              95               84
# Hymenoptera                 23                  101              98              116
# Lepidoptera                104                  261             239               99

##%######################################################%##
#                                                          #
####     2. Rerun the models removing these outliers    ####
#                                                          #
##%######################################################%##

# here the models are rerun after the outliers have been removed from the dataset

# reload data if required
datadir <- "5_RunLUIClimateModels/"
predictsSites <- readRDS(file = paste0(datadir,"PREDICTSSitesClimate_Data.rds"))



# 1. Abundance, mean anomaly

# list of studies that had larger Cook's distance values
abun_ss <- c("AD1_2008__Billeter 6",   "HW1_2011__Summerville 1", "SC1_2005__Richardson 1") # taken from above
outliers <- abun_ss

# subset the data to exclude the potential outliers
model_data <- predictsSites[!predictsSites$SS %in% outliers, ] # 8275 rows

model_data <- model_data[!is.na(model_data$LogAbund), ] # 7885 rows

length(unique(model_data$SS)) # 232

MeanAnomalyModelAbund <- GLMER(modelData = model_data, 
                               responseVar = "LogAbund",
                               fitFamily = "gaussian",
                               fixedStruct = "LUI * StdTmeanAnomalyRS * Order",
                               randomStruct = "(1|SS)+(1|SSB)",
                               saveVars = c("SSBS"))

# save the model output
save(MeanAnomalyModelAbund, file = paste0(outdir, "/MeanAnomalyModelAbund_rmout.rdata"))
#load(file = paste0(outdir, "/MeanAnomalyModelAbund_rmout.rdata"))



# 2. Richness, mean anomaly

# list of studies that had larger Cook's distance values
rich_ss <- c("SC1_2010__Marsh 1", "AD1_2008__Franzen 1", "SE1_2012__Poveda 1", 
             "LH1_2008__Littlewood 1",  "SE1_2012__Poveda 2", "HP1_2006__Lachat 1",
             "DI1_2013__Rousseau 2", "TN1_2007__Bouyer 2", "AD1_2008__Billeter 6",
             "SH1_2011__Todd 1", "SC1_2005__Richardson 1",  "SC1_2011__Meijer 1",     
             "AR1_2008__Basset 1", "HW1_2011__Summerville 1")
outliers <- rich_ss


# subset the data to exclude the potential outliers
model_data <- predictsSites[!predictsSites$SS %in% outliers, ] # 6651 rows

length(unique(model_data$SS)) # 239

MeanAnomalyModelRich <- GLMER(modelData = model_data,
                              responseVar = "Species_richness",
                              fitFamily = "poisson",
                              fixedStruct = "LUI * StdTmeanAnomalyRS * Order",
                              randomStruct = "(1|SS)+(1|SSB)+(1|SSBS)",
                              saveVars = c("SSBS"))

# save model output
save(MeanAnomalyModelRich, file = paste0(outdir, "/MeanAnomalyModelRich_rmout.rdata"))
#load(file = paste0(outdir, "/MeanAnomalyModelRich_rmout.rdata"))

MeanAnomalyModelRich$model



##### save model output tables ####

tab_model(MeanAnomalyModelAbund$model, MeanAnomalyModelRich$model, 
          show.ci = F, 
          show.se = T, 
          show.est = T, 
          transform = NULL, 
          show.icc = F, 
          show.re.var = F,
          file = paste0(outdir, "/Abun_Rich_outlierrm_output.html"))


R2GLMER(MeanAnomalyModelAbund$model)
# $conditional
# [1] 0.445493
# 
# $marginal
# [1] 0.1009416

R2GLMER(MeanAnomalyModelRich$model) # tab_model function gives incorrect values for the richness models
# $conditional
# [1] 0.7189628
# 
# $marginal
# [1] 0.05395956


##%######################################################%##
#                                                          #
####                3. Replot Figures                   ####
#                                                          #
##%######################################################%##


# set quantiles of predicted result to be presented in the plots
exclQuantiles <- c(0.025,0.975)

####  Abundance, Mean Anomaly ####

nd <- expand.grid(
  StdTmeanAnomalyRS=seq(from = min(MeanAnomalyModelAbund$data$StdTmeanAnomalyRS),
                        to = max(MeanAnomalyModelAbund$data$StdTmeanAnomalyRS),
                        length.out = 100),
  LUI=factor(c("Primary vegetation","Secondary vegetation","Agriculture_Low","Agriculture_High"),
             levels = levels(MeanAnomalyModelAbund$data$LUI)),
  Order=factor(c("Coleoptera","Diptera","Hemiptera","Hymenoptera","Lepidoptera")))

# back transform the predictors
nd$StdTmeanAnomaly <- BackTransformCentreredPredictor(
  transformedX = nd$StdTmeanAnomalyRS,
  originalX = predictsSites$StdTmeanAnomaly)

# set richness and abundance to 0 - to be predicted
nd$LogAbund <- 0
nd$Species_richness <- 0

# reference for % difference = primary vegetation and positive anomaly closest to 0
refRow <- which((nd$LUI=="Primary vegetation") & (nd$StdTmeanAnomaly==min(abs(nd$StdTmeanAnomaly))))

# set quantiles by Order
# coleoptera
C_QPV <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Primary vegetation" & MeanAnomalyModelAbund$data$Order == "Coleoptera"],
  probs = exclQuantiles)
C_QSV <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Secondary vegetation" & MeanAnomalyModelAbund$data$Order == "Coleoptera"],
  probs = exclQuantiles)
C_QAL <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Agriculture_Low" & MeanAnomalyModelAbund$data$Order == "Coleoptera"],
  probs = exclQuantiles)
C_QAH <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Agriculture_High" & MeanAnomalyModelAbund$data$Order == "Coleoptera"],
  probs = exclQuantiles)

# diptera
D_QPV <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Primary vegetation" & MeanAnomalyModelAbund$data$Order == "Diptera"],
  probs = exclQuantiles)
D_QSV <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Secondary vegetation" & MeanAnomalyModelAbund$data$Order == "Diptera"],
  probs = exclQuantiles)
D_QAL <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Agriculture_Low" & MeanAnomalyModelAbund$data$Order == "Diptera"],
  probs = exclQuantiles)
D_QAH <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Agriculture_High" & MeanAnomalyModelAbund$data$Order == "Diptera"],
  probs = exclQuantiles)

# hemiptera
He_QPV <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Primary vegetation" & MeanAnomalyModelAbund$data$Order == "Hemiptera"],
  probs = exclQuantiles)
He_QSV <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Secondary vegetation" & MeanAnomalyModelAbund$data$Order == "Hemiptera"],
  probs = exclQuantiles)
He_QAL <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Agriculture_Low" & MeanAnomalyModelAbund$data$Order == "Hemiptera"],
  probs = exclQuantiles)
He_QAH <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Agriculture_High" & MeanAnomalyModelAbund$data$Order == "Hemiptera"],
  probs = exclQuantiles)

# hymenoptera
Hy_QPV <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Primary vegetation" & MeanAnomalyModelAbund$data$Order == "Hymenoptera"],
  probs = exclQuantiles)
Hy_QSV <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Secondary vegetation" & MeanAnomalyModelAbund$data$Order == "Hymenoptera"],
  probs = exclQuantiles)
Hy_QAL <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Agriculture_Low" & MeanAnomalyModelAbund$data$Order == "Hymenoptera"],
  probs = exclQuantiles)
Hy_QAH <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Agriculture_High" & MeanAnomalyModelAbund$data$Order == "Hymenoptera"],
  probs = exclQuantiles)

# lepidoptera
L_QPV <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Primary vegetation" & MeanAnomalyModelAbund$data$Order == "Lepidoptera"],
  probs = exclQuantiles)
L_QSV <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Secondary vegetation" & MeanAnomalyModelAbund$data$Order == "Lepidoptera"],
  probs = exclQuantiles)
L_QAL <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Agriculture_Low" & MeanAnomalyModelAbund$data$Order == "Lepidoptera"],
  probs = exclQuantiles)
L_QAH <- quantile(x = MeanAnomalyModelAbund$data$StdTmeanAnomalyRS[
  MeanAnomalyModelAbund$data$LUI=="Agriculture_High" & MeanAnomalyModelAbund$data$Order == "Lepidoptera"],
  probs = exclQuantiles)


# predict the results
a.preds.tmean <- PredictGLMERRandIter(model = MeanAnomalyModelAbund$model,data = nd)

# back transform the abundance values
a.preds.tmean <- exp(a.preds.tmean)-0.01

# create list of matrices
number_of_chunks = 5
list_a.preds.tmean <- lapply(seq(1, NROW(a.preds.tmean), ceiling(NROW(a.preds.tmean)/number_of_chunks)),
                             function(i) a.preds.tmean[i:min(i + ceiling(NROW(a.preds.tmean)/number_of_chunks) - 1, NROW(a.preds.tmean)),])

# name them
names(list_a.preds.tmean) <- c("Coleoptera","Diptera","Hemiptera","Hymenoptera","Lepidoptera")

# sweep out according to the refRow
list_a.preds.tmean <- lapply(list_a.preds.tmean,FUN=function(x){
  sweep (x=x, MARGIN = 2, STATS=x[refRow[1],],FUN="/") 
})


# print to global environment
list2env(list_a.preds.tmean,globalenv())

# split nd by order
Order<- paste0("nd_",nd$Order)
# create a list of data frames
by_Order <- split(nd,Order)
list2env(by_Order,globalenv())

# remove anything above and below the quantiles
Coleoptera[which((nd_Coleoptera$LUI=="Primary vegetation") & (nd_Coleoptera$StdTmeanAnomalyRS < C_QPV[1])),] <- NA
Coleoptera[which((nd_Coleoptera$LUI=="Primary vegetation") & (nd_Coleoptera$StdTmeanAnomalyRS > C_QPV[2])),] <- NA
Coleoptera[which((nd_Coleoptera$LUI=="Secondary vegetation") & (nd_Coleoptera$StdTmeanAnomalyRS < C_QSV[1])),] <- NA
Coleoptera[which((nd_Coleoptera$LUI=="Secondary vegetation") & (nd_Coleoptera$StdTmeanAnomalyRS > C_QSV[2])),] <- NA
Coleoptera[which((nd_Coleoptera$LUI=="Agriculture_Low") & (nd_Coleoptera$StdTmeanAnomalyRS < C_QAL[1])),] <- NA
Coleoptera[which((nd_Coleoptera$LUI=="Agriculture_Low") & (nd_Coleoptera$StdTmeanAnomalyRS > C_QAL[2])),] <- NA
Coleoptera[which((nd_Coleoptera$LUI=="Agriculture_High") & (nd_Coleoptera$StdTmeanAnomalyRS < C_QAH[1])),] <- NA
Coleoptera[which((nd_Coleoptera$LUI=="Agriculture_High") & (nd_Coleoptera$StdTmeanAnomalyRS > C_QAH[2])),] <- NA

Diptera[which((nd_Diptera$LUI=="Primary vegetation") & (nd_Diptera$StdTmeanAnomalyRS < D_QPV[1])),] <- NA
Diptera[which((nd_Diptera$LUI=="Primary vegetation") & (nd_Diptera$StdTmeanAnomalyRS > D_QPV[2])),] <- NA
Diptera[which((nd_Diptera$LUI=="Secondary vegetation") & (nd_Diptera$StdTmeanAnomalyRS < D_QSV[1])),] <- NA
Diptera[which((nd_Diptera$LUI=="Secondary vegetation") & (nd_Diptera$StdTmeanAnomalyRS > D_QSV[2])),] <- NA
Diptera[which((nd_Diptera$LUI=="Agriculture_Low") & (nd_Diptera$StdTmeanAnomalyRS < D_QAL[1])),] <- NA
Diptera[which((nd_Diptera$LUI=="Agriculture_Low") & (nd_Diptera$StdTmeanAnomalyRS > D_QAL[2])),] <- NA
Diptera[which((nd_Diptera$LUI=="Agriculture_High") & (nd_Diptera$StdTmeanAnomalyRS < D_QAH[1])),] <- NA
Diptera[which((nd_Diptera$LUI=="Agriculture_High") & (nd_Diptera$StdTmeanAnomalyRS > D_QAH[2])),] <- NA

Hemiptera[which((nd_Hemiptera$LUI=="Primary vegetation") & (nd_Hemiptera$StdTmeanAnomalyRS < He_QPV[1])),] <- NA
Hemiptera[which((nd_Hemiptera$LUI=="Primary vegetation") & (nd_Hemiptera$StdTmeanAnomalyRS > He_QPV[2])),] <- NA
Hemiptera[which((nd_Hemiptera$LUI=="Secondary vegetation") & (nd_Hemiptera$StdTmeanAnomalyRS < He_QSV[1])),] <- NA
Hemiptera[which((nd_Hemiptera$LUI=="Secondary vegetation") & (nd_Hemiptera$StdTmeanAnomalyRS > He_QSV[2])),] <- NA
Hemiptera[which((nd_Hemiptera$LUI=="Agriculture_Low") & (nd_Hemiptera$StdTmeanAnomalyRS < He_QAL[1])),] <- NA
Hemiptera[which((nd_Hemiptera$LUI=="Agriculture_Low") & (nd_Hemiptera$StdTmeanAnomalyRS > He_QAL[2])),] <- NA
Hemiptera[which((nd_Hemiptera$LUI=="Agriculture_High") & (nd_Hemiptera$StdTmeanAnomalyRS < He_QAH[1])),] <- NA
Hemiptera[which((nd_Hemiptera$LUI=="Agriculture_High") & (nd_Hemiptera$StdTmeanAnomalyRS > He_QAH[2])),] <- NA

Hymenoptera[which((nd_Hymenoptera$LUI=="Primary vegetation") & (nd_Hymenoptera$StdTmeanAnomalyRS < Hy_QPV[1])),] <- NA
Hymenoptera[which((nd_Hymenoptera$LUI=="Primary vegetation") & (nd_Hymenoptera$StdTmeanAnomalyRS > Hy_QPV[2])),] <- NA
Hymenoptera[which((nd_Hymenoptera$LUI=="Secondary vegetation") & (nd_Hymenoptera$StdTmeanAnomalyRS < Hy_QSV[1])),] <- NA
Hymenoptera[which((nd_Hymenoptera$LUI=="Secondary vegetation") & (nd_Hymenoptera$StdTmeanAnomalyRS > Hy_QSV[2])),] <- NA
Hymenoptera[which((nd_Hymenoptera$LUI=="Agriculture_Low") & (nd_Hymenoptera$StdTmeanAnomalyRS < Hy_QAL[1])),] <- NA
Hymenoptera[which((nd_Hymenoptera$LUI=="Agriculture_Low") & (nd_Hymenoptera$StdTmeanAnomalyRS > Hy_QAL[2])),] <- NA
Hymenoptera[which((nd_Hymenoptera$LUI=="Agriculture_High") & (nd_Hymenoptera$StdTmeanAnomalyRS < Hy_QAH[1])),] <- NA
Hymenoptera[which((nd_Hymenoptera$LUI=="Agriculture_High") & (nd_Hymenoptera$StdTmeanAnomalyRS > Hy_QAH[2])),] <- NA

Lepidoptera[which((nd_Lepidoptera$LUI=="Primary vegetation") & (nd_Lepidoptera$StdTmeanAnomalyRS < L_QPV[1])),] <- NA
Lepidoptera[which((nd_Lepidoptera$LUI=="Primary vegetation") & (nd_Lepidoptera$StdTmeanAnomalyRS > L_QPV[2])),] <- NA
Lepidoptera[which((nd_Lepidoptera$LUI=="Secondary vegetation") & (nd_Lepidoptera$StdTmeanAnomalyRS < L_QSV[1])),] <- NA
Lepidoptera[which((nd_Lepidoptera$LUI=="Secondary vegetation") & (nd_Lepidoptera$StdTmeanAnomalyRS > L_QSV[2])),] <- NA
Lepidoptera[which((nd_Lepidoptera$LUI=="Agriculture_Low") & (nd_Lepidoptera$StdTmeanAnomalyRS < L_QAL[1])),] <- NA
Lepidoptera[which((nd_Lepidoptera$LUI=="Agriculture_Low") & (nd_Lepidoptera$StdTmeanAnomalyRS > L_QAL[2])),] <- NA
Lepidoptera[which((nd_Lepidoptera$LUI=="Agriculture_High") & (nd_Lepidoptera$StdTmeanAnomalyRS < L_QAH[1])),] <- NA
Lepidoptera[which((nd_Lepidoptera$LUI=="Agriculture_High") & (nd_Lepidoptera$StdTmeanAnomalyRS > L_QAH[2])),] <- NA


# Get the median, upper and lower quants for the plot

nd_Coleoptera$PredMedian <- ((apply(X = Coleoptera,MARGIN = 1,
                                    FUN = median,na.rm=TRUE))*100)-100
nd_Coleoptera$PredUpper <- ((apply(X = Coleoptera,MARGIN = 1,
                                   FUN = quantile,probs = 0.975,na.rm=TRUE))*100)-100
nd_Coleoptera$PredLower <- ((apply(X = Coleoptera,MARGIN = 1,
                                   FUN = quantile,probs = 0.025,na.rm=TRUE))*100)-100

nd_Diptera$PredMedian <- ((apply(X = Diptera,MARGIN = 1,
                                 FUN = median,na.rm=TRUE))*100)-100
nd_Diptera$PredUpper <- ((apply(X = Diptera,MARGIN = 1,
                                FUN = quantile,probs = 0.975,na.rm=TRUE))*100)-100
nd_Diptera$PredLower <- ((apply(X = Diptera,MARGIN = 1,
                                FUN = quantile,probs = 0.025,na.rm=TRUE))*100)-100

nd_Hemiptera$PredMedian <- ((apply(X = Hemiptera,MARGIN = 1,
                                   FUN = median,na.rm=TRUE))*100)-100
nd_Hemiptera$PredUpper <- ((apply(X = Hemiptera,MARGIN = 1,
                                  FUN = quantile,probs = 0.975,na.rm=TRUE))*100)-100
nd_Hemiptera$PredLower <- ((apply(X = Hemiptera,MARGIN = 1,
                                  FUN = quantile,probs = 0.025,na.rm=TRUE))*100)-100

nd_Hymenoptera$PredMedian <- ((apply(X = Hymenoptera,MARGIN = 1,
                                     FUN = median,na.rm=TRUE))*100)-100
nd_Hymenoptera$PredUpper <- ((apply(X = Hymenoptera,MARGIN = 1,
                                    FUN = quantile,probs = 0.975,na.rm=TRUE))*100)-100
nd_Hymenoptera$PredLower <- ((apply(X = Hymenoptera,MARGIN = 1,
                                    FUN = quantile,probs = 0.025,na.rm=TRUE))*100)-100

nd_Lepidoptera$PredMedian <- ((apply(X = Lepidoptera,MARGIN = 1,
                                     FUN = median,na.rm=TRUE))*100)-100
nd_Lepidoptera$PredUpper <- ((apply(X = Lepidoptera,MARGIN = 1,
                                    FUN = quantile,probs = 0.975,na.rm=TRUE))*100)-100
nd_Lepidoptera$PredLower <- ((apply(X = Lepidoptera,MARGIN = 1,
                                    FUN = quantile,probs = 0.025,na.rm=TRUE))*100)-100


# set factor levels
nd_Coleoptera$LUI <- factor(nd_Coleoptera$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))
nd_Diptera$LUI <- factor(nd_Diptera$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))
nd_Hemiptera$LUI <- factor(nd_Hemiptera$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))
nd_Hymenoptera$LUI <- factor(nd_Hymenoptera$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))
nd_Lepidoptera$LUI <- factor(nd_Lepidoptera$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))

# plot

p_coleoptera <- ggplot(data = nd_Coleoptera, aes(x = StdTmeanAnomaly, y = PredMedian)) + 
  geom_line(aes(col = LUI), size = 0.75) +
  geom_ribbon(aes(ymin = nd_Coleoptera$PredLower, ymax = nd_Coleoptera$PredUpper, fill = LUI), alpha = 0.2) +
  geom_hline(yintercept = 0, lty = "dashed", size = 0.25) +
  scale_fill_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  scale_colour_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  theme_bw() + 
  scale_x_continuous(breaks = c(0,0.5, 1, 1.5, 2), limits = c(0, 2)) +
  #scale_x_continuous(breaks = c(0, 0.25, 0.5, 0.75, 1), limits = c(0, 1)) +
  scale_y_continuous(breaks = c(-100,-75, -50, -25, 0, 25, 50, 75, 100, 125,150,175), limits = c(-100, 175)) +
  ylab("Change in total abundance (%)") +
  xlab("Standardised Temperature Anomaly") + 
  ggtitle("Coleoptera") +
  theme(aspect.ratio = 1, 
        plot.title = element_text(size = 8, face = 'bold'),
        axis.title.x = element_text(size = 8),
        axis.title.y = element_text(size = 8),
        axis.text = element_text(size = 7),
        axis.ticks = element_line(size = 0.25),
        legend.position = "none", 
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(size = 0.25),
        panel.border = element_rect(size = 0.25)) 


p_diptera <- ggplot(data = nd_Diptera, aes(x = StdTmeanAnomaly, y = PredMedian)) + 
  geom_line(aes(col = LUI), size = 0.75) +
  geom_ribbon(aes(ymin = nd_Diptera$PredLower, ymax = nd_Diptera$PredUpper, fill = LUI), alpha = 0.2) +
  geom_hline(yintercept = 0, lty = "dashed", size = 0.25) +
  scale_fill_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  scale_colour_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  theme_bw() + 
  scale_x_continuous(breaks = c(0,0.5, 1, 1.5, 2), limits = c(0, 2)) +
  #scale_x_continuous(breaks = c(0, 0.25, 0.5, 0.75, 1), limits = c(0, 1)) +
  scale_y_continuous(breaks = c(-100,-75, -50, -25, 0, 25, 50, 75, 100, 125,150,175), limits = c(-100, 175)) +
  ylab("Change in total abundance (%)") +
  xlab("Standardised Temperature Anomaly") +
  ggtitle("Diptera") +
  theme(aspect.ratio = 1, 
        plot.title = element_text(size = 8, face = 'bold'),
        axis.title = element_text(size = 8),
        axis.text = element_text(size = 7),
        axis.ticks = element_line(size = 0.25),
        legend.position = "none", 
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(size = 0.25),
        panel.border = element_rect(size = 0.25)) 


p_hemiptera <- ggplot(data = nd_Hemiptera, aes(x = StdTmeanAnomaly, y = PredMedian)) + 
  geom_line(aes(col = LUI), size = 0.75) +
  geom_ribbon(aes(ymin = nd_Hemiptera$PredLower, ymax = nd_Hemiptera$PredUpper, fill = LUI), alpha = 0.2) +
  geom_hline(yintercept = 0, lty = "dashed", size = 0.25) +
  scale_fill_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  scale_colour_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  theme_bw() + 
  scale_x_continuous(breaks = c(0,0.5, 1, 1.5, 2), limits = c(0, 2)) +
  #scale_x_continuous(breaks = c(0, 0.25, 0.5, 0.75, 1), limits = c(0, 1)) +
  scale_y_continuous(breaks = c(-100,-75, -50, -25, 0, 25, 50, 75, 100, 125,150,175), limits = c(-100, 175)) +
  ylab("Change in total abundance (%)") +
  xlab("Standardised Temperature Anomaly") +
  ggtitle("Hemiptera") +
  theme(aspect.ratio = 1, 
        plot.title = element_text(size = 8, face = 'bold'),
        axis.title = element_text(size = 8),
        axis.text = element_text(size = 7),
        axis.ticks = element_line(size = 0.25),
        legend.position = "none",
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(size = 0.25),
        panel.border = element_rect(size = 0.25))


p_hymenoptera <- ggplot(data = nd_Hymenoptera, aes(x = StdTmeanAnomaly, y = PredMedian)) + 
  geom_line(aes(col = LUI), size = 0.75) +
  geom_ribbon(aes(ymin = nd_Hymenoptera$PredLower, ymax = nd_Hymenoptera$PredUpper, fill = LUI), alpha = 0.2) +
  geom_hline(yintercept = 0, lty = "dashed", size = 0.25) +
  scale_fill_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  scale_colour_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  theme_bw() + 
  scale_x_continuous(breaks = c(0,0.5, 1, 1.5, 2), limits = c(0, 2)) +
  #scale_x_continuous(breaks = c(0, 0.25, 0.5, 0.75, 1), limits = c(0, 1)) +
  scale_y_continuous(breaks = c(-100,-75, -50, -25, 0, 25, 50, 75, 100, 125,150,175), limits = c(-100, 225)) +
  ylab("Change in total abundance (%)") +
  xlab("Standardised Temperature Anomaly") +
  ggtitle("Hymenoptera") +
  theme(aspect.ratio = 1, 
        plot.title = element_text(size = 8, face = 'bold'),
        axis.title = element_text(size = 8),
        axis.text = element_text(size = 7),
        axis.ticks = element_line(size = 0.25),
        legend.position = "none", 
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(size = 0.25),
        panel.border = element_rect(size = 0.25))

p_lepidoptera <- ggplot(data = nd_Lepidoptera, aes(x = StdTmeanAnomaly, y = PredMedian)) + 
  geom_line(aes(col = LUI), size = 0.75) +
  geom_ribbon(aes(ymin = nd_Lepidoptera$PredLower, ymax = nd_Lepidoptera$PredUpper, fill = LUI), alpha = 0.2) +
  geom_hline(yintercept = 0, lty = "dashed", size = 0.25) +
  scale_fill_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  scale_colour_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  theme_bw() + 
  scale_x_continuous(breaks = c(0,0.5, 1, 1.5, 2), limits = c(0, 2)) +
  #scale_x_continuous(breaks = c(0, 0.25, 0.5, 0.75, 1), limits = c(0, 1)) +
  scale_y_continuous(breaks = c(-100,-75, -50, -25, 0, 25, 50, 75, 100, 125,150,175), limits = c(-100, 175)) +
  ylab("Change in total abundance (%)") +
  xlab("Standardised Temperature Anomaly") +
  ggtitle("Lepidoptera") +
  theme(aspect.ratio = 1, 
        plot.title = element_text(size = 8, face = 'bold'),
        axis.title = element_text(size = 8),
        axis.text = element_text(size = 7),
        axis.ticks = element_line(size = 0.25),
        legend.position = "none",
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(size = 0.25),
        panel.border = element_rect(size = 0.25)) 


# get the legend
legend <- get_legend(
  p_coleoptera +
    guides(color = guide_legend(ncol = 1)) +
    theme(legend.position = "right",
          legend.background = element_blank(), 
          legend.text = element_text(size = 7), 
          legend.title = element_blank())
)


# put them all together to save them
MeanAnomAbund <- cowplot::plot_grid(p_coleoptera,p_diptera,p_hemiptera,p_hymenoptera,p_lepidoptera,legend, ncol=3)

# save plot (pdf)
ggsave(filename = paste0(outdir, "MeanAnomAbund_outrm.pdf"), plot = MeanAnomAbund, width = 200, height = 150, units = "mm", dpi = 300)

# save plot (jpeg)
ggsave("MeanAnomAbund_outrm.jpeg", device ="jpeg", path = outdir, width=20, height=15, units="cm", dpi = 350)


#### Richness, Mean Anomaly ####

nd2 <- expand.grid(
  StdTmeanAnomalyRS=seq(from = min(MeanAnomalyModelRich$data$StdTmeanAnomalyRS),
                        to = max(MeanAnomalyModelRich$data$StdTmeanAnomalyRS),
                        length.out = 100),
  LUI=factor(c("Primary vegetation","Secondary vegetation","Agriculture_Low","Agriculture_High"),
             levels = levels(MeanAnomalyModelRich$data$LUI)),
  Order=factor(c("Coleoptera","Diptera","Hemiptera","Hymenoptera","Lepidoptera")))

# back transform the predictors
nd2$StdTmeanAnomaly <- BackTransformCentreredPredictor(
  transformedX = nd2$StdTmeanAnomalyRS,
  originalX = predictsSites$StdTmeanAnomaly)

# set richness and abundance to 0 - to be predicted
nd2$LogAbund <- 0
nd2$Species_richness <- 0

# reference for % difference = primary vegetation and positive anomaly closest to 0
refRow <- which((nd2$LUI=="Primary vegetation") & (nd2$StdTmeanAnomaly==min(abs(nd2$StdTmeanAnomaly))))

# set quantiles
# coleoptera
C_QPV <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Primary vegetation" & MeanAnomalyModelRich$data$Order == "Coleoptera"],
  probs = exclQuantiles)
C_QSV <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Secondary vegetation" & MeanAnomalyModelRich$data$Order == "Coleoptera"],
  probs = exclQuantiles)
C_QAL <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Agriculture_Low" & MeanAnomalyModelRich$data$Order == "Coleoptera"],
  probs = exclQuantiles)
C_QAH <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Agriculture_High" & MeanAnomalyModelRich$data$Order == "Coleoptera"],
  probs = exclQuantiles)

# diptera
D_QPV <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Primary vegetation" & MeanAnomalyModelRich$data$Order == "Diptera"],
  probs = exclQuantiles)
D_QSV <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Secondary vegetation" & MeanAnomalyModelRich$data$Order == "Diptera"],
  probs = exclQuantiles)
D_QAL <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Agriculture_Low" & MeanAnomalyModelRich$data$Order == "Diptera"],
  probs = exclQuantiles)
D_QAH <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Agriculture_High" & MeanAnomalyModelRich$data$Order == "Diptera"],
  probs = exclQuantiles)

# hemiptera
He_QPV <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Primary vegetation" & MeanAnomalyModelRich$data$Order == "Hemiptera"],
  probs = exclQuantiles)
He_QSV <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Secondary vegetation" & MeanAnomalyModelRich$data$Order == "Hemiptera"],
  probs = exclQuantiles)
He_QAL <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Agriculture_Low" & MeanAnomalyModelRich$data$Order == "Hemiptera"],
  probs = exclQuantiles)
He_QAH <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Agriculture_High" & MeanAnomalyModelRich$data$Order == "Hemiptera"],
  probs = exclQuantiles)

# hymenoptera
Hy_QPV <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Primary vegetation" & MeanAnomalyModelRich$data$Order == "Hymenoptera"],
  probs = exclQuantiles)
Hy_QSV <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Secondary vegetation" & MeanAnomalyModelRich$data$Order == "Hymenoptera"],
  probs = exclQuantiles)
Hy_QAL <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Agriculture_Low" & MeanAnomalyModelRich$data$Order == "Hymenoptera"],
  probs = exclQuantiles)
Hy_QAH <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Agriculture_High" & MeanAnomalyModelRich$data$Order == "Hymenoptera"],
  probs = exclQuantiles)

# lepidoptera
L_QPV <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Primary vegetation" & MeanAnomalyModelRich$data$Order == "Lepidoptera"],
  probs = exclQuantiles)
L_QSV <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Secondary vegetation" & MeanAnomalyModelRich$data$Order == "Lepidoptera"],
  probs = exclQuantiles)
L_QAL <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Agriculture_Low" & MeanAnomalyModelRich$data$Order == "Lepidoptera"],
  probs = exclQuantiles)
L_QAH <- quantile(x = MeanAnomalyModelRich$data$StdTmeanAnomalyRS[
  MeanAnomalyModelRich$data$LUI=="Agriculture_High" & MeanAnomalyModelRich$data$Order == "Lepidoptera"],
  probs = exclQuantiles)


# predict the results
sr.preds.tmean <- PredictGLMERRandIter(model = MeanAnomalyModelRich$model,data = nd2)

# back transform the abundance values
sr.preds.tmean <- exp(sr.preds.tmean)-0.01

# create list of matrices
number_of_chunks = 5
list_sr.preds.tmean <- lapply(seq(1, NROW(sr.preds.tmean), ceiling(NROW(sr.preds.tmean)/number_of_chunks)),
                              function(i) sr.preds.tmean[i:min(i + ceiling(NROW(sr.preds.tmean)/number_of_chunks) - 1, NROW(sr.preds.tmean)),])
# name them
names(list_sr.preds.tmean) <- c("Coleoptera","Diptera","Hemiptera","Hymenoptera","Lepidoptera")

# sweep out according to the refRow
list_sr.preds.tmean <- lapply(list_sr.preds.tmean,FUN=function(x){
  sweep (x=x, MARGIN = 2, STATS=x[refRow[1],],FUN="/") 
})

list2env(list_sr.preds.tmean,globalenv())

# split nd by order
Order<- paste0("nd2_",nd2$Order)
# create a list of data frames
by_Order <- split(nd2,Order)
list2env(by_Order,globalenv())

# remove anything above and below the quantiles
Coleoptera[which((nd2_Coleoptera$LUI=="Primary vegetation") & (nd2_Coleoptera$StdTmeanAnomalyRS < C_QPV[1])),] <- NA
Coleoptera[which((nd2_Coleoptera$LUI=="Primary vegetation") & (nd2_Coleoptera$StdTmeanAnomalyRS > C_QPV[2])),] <- NA
Coleoptera[which((nd2_Coleoptera$LUI=="Secondary vegetation") & (nd2_Coleoptera$StdTmeanAnomalyRS < C_QSV[1])),] <- NA
Coleoptera[which((nd2_Coleoptera$LUI=="Secondary vegetation") & (nd2_Coleoptera$StdTmeanAnomalyRS > C_QSV[2])),] <- NA
Coleoptera[which((nd2_Coleoptera$LUI=="Agriculture_Low") & (nd2_Coleoptera$StdTmeanAnomalyRS < C_QAL[1])),] <- NA
Coleoptera[which((nd2_Coleoptera$LUI=="Agriculture_Low") & (nd2_Coleoptera$StdTmeanAnomalyRS > C_QAL[2])),] <- NA
Coleoptera[which((nd2_Coleoptera$LUI=="Agriculture_High") & (nd2_Coleoptera$StdTmeanAnomalyRS < C_QAH[1])),] <- NA
Coleoptera[which((nd2_Coleoptera$LUI=="Agriculture_High") & (nd2_Coleoptera$StdTmeanAnomalyRS > C_QAH[2])),] <- NA

Diptera[which((nd2_Diptera$LUI=="Primary vegetation") & (nd2_Diptera$StdTmeanAnomalyRS < D_QPV[1])),] <- NA
Diptera[which((nd2_Diptera$LUI=="Primary vegetation") & (nd2_Diptera$StdTmeanAnomalyRS > D_QPV[2])),] <- NA
Diptera[which((nd2_Diptera$LUI=="Secondary vegetation") & (nd2_Diptera$StdTmeanAnomalyRS < D_QSV[1])),] <- NA
Diptera[which((nd2_Diptera$LUI=="Secondary vegetation") & (nd2_Diptera$StdTmeanAnomalyRS > D_QSV[2])),] <- NA
Diptera[which((nd2_Diptera$LUI=="Agriculture_Low") & (nd2_Diptera$StdTmeanAnomalyRS < D_QAL[1])),] <- NA
Diptera[which((nd2_Diptera$LUI=="Agriculture_Low") & (nd2_Diptera$StdTmeanAnomalyRS > D_QAL[2])),] <- NA
Diptera[which((nd2_Diptera$LUI=="Agriculture_High") & (nd2_Diptera$StdTmeanAnomalyRS < D_QAH[1])),] <- NA
Diptera[which((nd2_Diptera$LUI=="Agriculture_High") & (nd2_Diptera$StdTmeanAnomalyRS > D_QAH[2])),] <- NA

Hemiptera[which((nd2_Hemiptera$LUI=="Primary vegetation") & (nd2_Hemiptera$StdTmeanAnomalyRS < He_QPV[1])),] <- NA
Hemiptera[which((nd2_Hemiptera$LUI=="Primary vegetation") & (nd2_Hemiptera$StdTmeanAnomalyRS > He_QPV[2])),] <- NA
Hemiptera[which((nd2_Hemiptera$LUI=="Secondary vegetation") & (nd2_Hemiptera$StdTmeanAnomalyRS < He_QSV[1])),] <- NA
Hemiptera[which((nd2_Hemiptera$LUI=="Secondary vegetation") & (nd2_Hemiptera$StdTmeanAnomalyRS > He_QSV[2])),] <- NA
Hemiptera[which((nd2_Hemiptera$LUI=="Agriculture_Low") & (nd2_Hemiptera$StdTmeanAnomalyRS < He_QAL[1])),] <- NA
Hemiptera[which((nd2_Hemiptera$LUI=="Agriculture_Low") & (nd2_Hemiptera$StdTmeanAnomalyRS > He_QAL[2])),] <- NA
Hemiptera[which((nd2_Hemiptera$LUI=="Agriculture_High") & (nd2_Hemiptera$StdTmeanAnomalyRS < He_QAH[1])),] <- NA
Hemiptera[which((nd2_Hemiptera$LUI=="Agriculture_High") & (nd2_Hemiptera$StdTmeanAnomalyRS > He_QAH[2])),] <- NA

Hymenoptera[which((nd2_Hymenoptera$LUI=="Primary vegetation") & (nd2_Hymenoptera$StdTmeanAnomalyRS < Hy_QPV[1])),] <- NA
Hymenoptera[which((nd2_Hymenoptera$LUI=="Primary vegetation") & (nd2_Hymenoptera$StdTmeanAnomalyRS > Hy_QPV[2])),] <- NA
Hymenoptera[which((nd2_Hymenoptera$LUI=="Secondary vegetation") & (nd2_Hymenoptera$StdTmeanAnomalyRS < Hy_QSV[1])),] <- NA
Hymenoptera[which((nd2_Hymenoptera$LUI=="Secondary vegetation") & (nd2_Hymenoptera$StdTmeanAnomalyRS > Hy_QSV[2])),] <- NA
Hymenoptera[which((nd2_Hymenoptera$LUI=="Agriculture_Low") & (nd2_Hymenoptera$StdTmeanAnomalyRS < Hy_QAL[1])),] <- NA
Hymenoptera[which((nd2_Hymenoptera$LUI=="Agriculture_Low") & (nd2_Hymenoptera$StdTmeanAnomalyRS > Hy_QAL[2])),] <- NA
Hymenoptera[which((nd2_Hymenoptera$LUI=="Agriculture_High") & (nd2_Hymenoptera$StdTmeanAnomalyRS < Hy_QAH[1])),] <- NA
Hymenoptera[which((nd2_Hymenoptera$LUI=="Agriculture_High") & (nd2_Hymenoptera$StdTmeanAnomalyRS > Hy_QAH[2])),] <- NA

Lepidoptera[which((nd2_Lepidoptera$LUI=="Primary vegetation") & (nd2_Lepidoptera$StdTmeanAnomalyRS < L_QPV[1])),] <- NA
Lepidoptera[which((nd2_Lepidoptera$LUI=="Primary vegetation") & (nd2_Lepidoptera$StdTmeanAnomalyRS > L_QPV[2])),] <- NA
Lepidoptera[which((nd2_Lepidoptera$LUI=="Secondary vegetation") & (nd2_Lepidoptera$StdTmeanAnomalyRS < L_QSV[1])),] <- NA
Lepidoptera[which((nd2_Lepidoptera$LUI=="Secondary vegetation") & (nd2_Lepidoptera$StdTmeanAnomalyRS > L_QSV[2])),] <- NA
Lepidoptera[which((nd2_Lepidoptera$LUI=="Agriculture_Low") & (nd2_Lepidoptera$StdTmeanAnomalyRS < L_QAL[1])),] <- NA
Lepidoptera[which((nd2_Lepidoptera$LUI=="Agriculture_Low") & (nd2_Lepidoptera$StdTmeanAnomalyRS > L_QAL[2])),] <- NA
Lepidoptera[which((nd2_Lepidoptera$LUI=="Agriculture_High") & (nd2_Lepidoptera$StdTmeanAnomalyRS < L_QAH[1])),] <- NA
Lepidoptera[which((nd2_Lepidoptera$LUI=="Agriculture_High") & (nd2_Lepidoptera$StdTmeanAnomalyRS > L_QAH[2])),] <- NA



# Get the median, upper and lower quants for the plot
nd2_Coleoptera$PredMedian <- ((apply(X = Coleoptera,MARGIN = 1,
                                     FUN = median,na.rm=TRUE))*100)-100
nd2_Coleoptera$PredUpper <- ((apply(X = Coleoptera,MARGIN = 1,
                                    FUN = quantile,probs = 0.975,na.rm=TRUE))*100)-100
nd2_Coleoptera$PredLower <- ((apply(X = Coleoptera,MARGIN = 1,
                                    FUN = quantile,probs = 0.025,na.rm=TRUE))*100)-100

nd2_Diptera$PredMedian <- ((apply(X = Diptera,MARGIN = 1,
                                  FUN = median,na.rm=TRUE))*100)-100
nd2_Diptera$PredUpper <- ((apply(X = Diptera,MARGIN = 1,
                                 FUN = quantile,probs = 0.975,na.rm=TRUE))*100)-100
nd2_Diptera$PredLower <- ((apply(X = Diptera,MARGIN = 1,
                                 FUN = quantile,probs = 0.025,na.rm=TRUE))*100)-100

nd2_Hemiptera$PredMedian <- ((apply(X = Hemiptera,MARGIN = 1,
                                    FUN = median,na.rm=TRUE))*100)-100
nd2_Hemiptera$PredUpper <- ((apply(X = Hemiptera,MARGIN = 1,
                                   FUN = quantile,probs = 0.975,na.rm=TRUE))*100)-100
nd2_Hemiptera$PredLower <- ((apply(X = Hemiptera,MARGIN = 1,
                                   FUN = quantile,probs = 0.025,na.rm=TRUE))*100)-100

nd2_Hymenoptera$PredMedian <- ((apply(X = Hymenoptera,MARGIN = 1,
                                      FUN = median,na.rm=TRUE))*100)-100
nd2_Hymenoptera$PredUpper <- ((apply(X = Hymenoptera,MARGIN = 1,
                                     FUN = quantile,probs = 0.975,na.rm=TRUE))*100)-100
nd2_Hymenoptera$PredLower <- ((apply(X = Hymenoptera,MARGIN = 1,
                                     FUN = quantile,probs = 0.025,na.rm=TRUE))*100)-100

nd2_Lepidoptera$PredMedian <- ((apply(X = Lepidoptera,MARGIN = 1,
                                      FUN = median,na.rm=TRUE))*100)-100
nd2_Lepidoptera$PredUpper <- ((apply(X = Lepidoptera,MARGIN = 1,
                                     FUN = quantile,probs = 0.975,na.rm=TRUE))*100)-100
nd2_Lepidoptera$PredLower <- ((apply(X = Lepidoptera,MARGIN = 1,
                                     FUN = quantile,probs = 0.025,na.rm=TRUE))*100)-100


# set factor levels
nd2_Coleoptera$LUI <- factor(nd2_Coleoptera$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))
nd2_Diptera$LUI <- factor(nd2_Diptera$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))
nd2_Hemiptera$LUI <- factor(nd2_Hemiptera$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))
nd2_Hymenoptera$LUI <- factor(nd2_Hymenoptera$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))
nd2_Lepidoptera$LUI <- factor(nd2_Lepidoptera$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))

# plot
p_coleoptera <- ggplot(data = nd2_Coleoptera, aes(x = StdTmeanAnomaly, y = PredMedian)) + 
  geom_line(aes(col = LUI), size = 0.75) +
  geom_ribbon(aes(ymin = nd2_Coleoptera$PredLower, ymax = nd2_Coleoptera$PredUpper, fill = LUI), alpha = 0.2) +
  geom_hline(yintercept = 0, lty = "dashed", size = 0.25) +
  scale_fill_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  scale_colour_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  theme_bw() + 
  scale_x_continuous(breaks = c(0,0.5, 1, 1.5, 2), limits = c(0, 2)) +
  scale_y_continuous(breaks = c(-100,-75, -50, -25, 0, 25, 50, 75, 100, 125,150,175), limits = c(-100, 175)) +
  ylab("Change in species richness (%)") +
  xlab("Standardised Temperature Anomaly") +
  ggtitle("Coleoptera") +
  theme(aspect.ratio = 1, 
        plot.title = element_text(size = 8, face = 'bold'),
        axis.title.x = element_text(size = 8),
        axis.title.y = element_text(size = 8),
        axis.text = element_text(size = 7),
        axis.ticks = element_line(size = 0.25),
        legend.position = "none",
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(size = 0.25),
        panel.border = element_rect(size = 0.25)) 

p_diptera <- ggplot(data = nd2_Diptera, aes(x = StdTmeanAnomaly, y = PredMedian)) + 
  geom_line(aes(col = LUI), size = 0.75) +
  geom_ribbon(aes(ymin = nd2_Diptera$PredLower, ymax = nd2_Diptera$PredUpper, fill = LUI), alpha = 0.2) +
  geom_hline(yintercept = 0, lty = "dashed", size = 0.25) +
  scale_fill_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  scale_colour_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  theme_bw() + 
  scale_x_continuous(breaks = c(0,0.5, 1, 1.5, 2), limits = c(0, 2)) +
  scale_y_continuous(breaks = c(-100,-75, -50, -25, 0, 25, 50, 75, 100, 125,150,175), limits = c(-100, 175)) +
  ylab("Change in species richness (%)") +
  xlab("Standardised Temperature Anomaly") +
  ggtitle("Diptera") +
  theme(aspect.ratio = 1, 
        plot.title = element_text(size = 8, face = 'bold'),
        axis.title = element_text(size = 8),
        axis.text = element_text(size = 7),
        axis.ticks = element_line(size = 0.25),
        legend.position = "none", 
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(size = 0.25),
        panel.border = element_rect(size = 0.25))

p_hemiptera <- ggplot(data = nd2_Hemiptera, aes(x = StdTmeanAnomaly, y = PredMedian)) + 
  geom_line(aes(col = LUI), size = 0.75) +
  geom_ribbon(aes(ymin = nd2_Hemiptera$PredLower, ymax = nd2_Hemiptera$PredUpper, fill = LUI), alpha = 0.2) +
  geom_hline(yintercept = 0, lty = "dashed", size = 0.25) +
  scale_fill_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  scale_colour_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  theme_bw() + 
  scale_x_continuous(breaks = c(0,0.5, 1, 1.5, 2), limits = c(0, 2)) +
  scale_y_continuous(breaks = c(-100,-75, -50, -25, 0, 25, 50, 75, 100, 125,150,175), limits = c(-100, 175)) +
  #scale_y_continuous(breaks = c(-100, 0, 100, 200, 300, 400, 500, 600, 700, 800, 900, 1000), limits = c(-100, 1000)) +
  ylab("Change in species richness (%)") +
  xlab("Standardised Temperature Anomaly") +
  ggtitle("Hemiptera") +
  theme(aspect.ratio = 1, 
        plot.title = element_text(size = 8, face = 'bold'),
        axis.title = element_text(size = 8),
        axis.text = element_text(size = 7),
        axis.ticks = element_line(size = 0.25),
        legend.position = "none",
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(size = 0.25),
        panel.border = element_rect(size = 0.25))

p_hymenoptera <- ggplot(data = nd2_Hymenoptera, aes(x = StdTmeanAnomaly, y = PredMedian)) + 
  geom_line(aes(col = LUI), size = 0.75) +
  geom_ribbon(aes(ymin = nd2_Hymenoptera$PredLower, ymax = nd2_Hymenoptera$PredUpper, fill = LUI), alpha = 0.2) +
  geom_hline(yintercept = 0, lty = "dashed", size = 0.25) +
  scale_fill_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  scale_colour_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  theme_bw() + 
  scale_x_continuous(breaks = c(0,0.5, 1, 1.5, 2), limits = c(0, 2)) +
  scale_y_continuous(breaks = c(-100,-75, -50, -25, 0, 25, 50, 75, 100, 125,150,175), limits = c(-100, 275)) +
  #scale_y_continuous(breaks = c(-100, 0, 100, 200, 300, 400, 500, 600, 700, 800, 900, 1000), limits = c(-100, 1000)) +
  ylab("Change in species richness (%)") +
  xlab("Standardised Temperature Anomaly") +
  ggtitle("Hymenoptera") +
  theme(aspect.ratio = 1, 
        plot.title = element_text(size = 8, face = 'bold'),
        axis.title = element_text(size = 8),
        axis.text = element_text(size = 7),
        axis.ticks = element_line(size = 0.25),
        legend.position = "none",
        #legend.position = c(0.2, 0.8),
        #legend.background = element_blank(), 
        #legend.text = element_text(size = 6), 
        #legend.title = element_blank(), 
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(size = 0.25),
        panel.border = element_rect(size = 0.25))

p_lepidoptera <- ggplot(data = nd2_Lepidoptera, aes(x = StdTmeanAnomaly, y = PredMedian)) + 
  geom_line(aes(col = LUI), size = 0.75) +
  geom_ribbon(aes(ymin = nd2_Lepidoptera$PredLower, ymax = nd2_Lepidoptera$PredUpper, fill = LUI), alpha = 0.2) +
  geom_hline(yintercept = 0, lty = "dashed", size = 0.25) +
  scale_fill_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  scale_colour_manual('Land-use', values = c("#009E73", "#0072B2","#E69F00","#D55E00")) +
  theme_bw() + 
  scale_x_continuous(breaks = c(0,0.5, 1, 1.5, 2), limits = c(0, 2)) +
  scale_y_continuous(breaks = c(-100,-75, -50, -25, 0, 25, 50, 75, 100, 125,150,175), limits = c(-100, 175)) +
  #scale_y_continuous(breaks = c(-100, 0, 100, 200, 300, 400, 500, 600, 700, 800, 900, 1000), limits = c(-100, 1000)) +
  ylab("Change in species richness (%)") +
  xlab("Standardised Temperature Anomaly") +
  ggtitle("Lepidoptera") +
  theme(aspect.ratio = 1, 
        plot.title = element_text(size = 8, face = 'bold'),
        axis.title = element_text(size = 8),
        axis.text = element_text(size = 7),
        axis.ticks = element_line(size = 0.25),
        legend.position = "none", 
        panel.grid.minor = element_blank(),
        panel.grid.major = element_line(size = 0.25),
        panel.border = element_rect(size = 0.25)) 

# get the legend
legend <- get_legend(
  p_coleoptera +
    guides(color = guide_legend(ncol = 1)) +
    theme(legend.position = "right",
          legend.background = element_blank(), 
          legend.text = element_text(size = 7), 
          legend.title = element_blank())
)


# put them all together to save them
MeanAnomRich <- cowplot::plot_grid(p_coleoptera,p_diptera,p_hemiptera,p_hymenoptera,p_lepidoptera,legend, ncol=3)

# save plot (pdf)
ggsave(filename = paste0(outdir, "MeanAnomRich_outrm.pdf"), plot = MeanAnomRich, width = 200, height = 150, units = "mm", dpi = 300)

# save plot (jpeg)
ggsave("MeanAnomRich_outrm.jpeg", device ="jpeg", path = outdir, width=20, height=15, units="cm", dpi = 350)



############################################################
#                                                          #
#                   Explore the outliers                   #
#                                                          #
############################################################

#  load in PREDICTS subset
datadir <- "5_RunLUIClimateModels/"
predictsSites <- readRDS(file = paste0(datadir,"PREDICTSSitesClimate_Data.rds"))

# load the models where outliers removed
load(file = paste0(outdir, "/MeanAnomalyModelAbund_rmout.rdata"))
load(file = paste0(outdir, "/MeanAnomalyModelRich_rmout.rdata"))

# determine list of outliers by comparing the datasets 
pred_abun <- predictsSites[!is.na(predictsSites$LogAbund), ]

ss_abun <- unique(pred_abun$SS[!pred_abun$SS %in% MeanAnomalyModelAbund$data$SS]) # 3
ss_rich <- unique(predictsSites$SS[!predictsSites$SS %in% MeanAnomalyModelRich$data$SS]) # 14

pred_about <- predictsSites[predictsSites$SS %in% ss_abun, ]
pred_about <- pred_about[!is.na(pred_about$LogAbund), ] # 583 rows
pred_srout <- predictsSites[predictsSites$SS %in% ss_rich, ] # 2207 rows



table(pred_about$LUI, pred_about$Order)
#                      Coleoptera Diptera Hemiptera Hymenoptera Lepidoptera
# Primary vegetation            0       0         0           0           0
# Agriculture_High              0      32         0          32           0
# Agriculture_Low               0      32         0          32           0
# Secondary vegetation        139      59        59          59         139


table(pred_srout$LUI, pred_srout$Order)
#                      Coleoptera Diptera Hemiptera Hymenoptera Lepidoptera
# Primary vegetation          101       0        78          23         104
# Agriculture_High             96     104        84         116          99
# Agriculture_Low             111      68        95          98         239
# Secondary vegetation        218      86       125         101         261


ss_abun %in% ss_rich # TRUE TRUE TRUE 
# all abundance outliers also richness outliers


#### Look at spread of anomaly values across outlier studies ####


library(ggplot2)
library(pals) # pals package has good palettes for discrete colours when large number of groups
library(ggfocus)

ggplot(data = pred_about, aes(x = StdTmeanAnomaly, y = LogAbund, col = SS)) + 
  geom_point() + 
  facet_wrap(~LUI) + 
  theme_bw()

ggsave(filename = paste0(outdir, "PLOT_Abun_outliers_STA_logAbun.png"))


ggplot(data = pred_srout, aes(x = StdTmeanAnomaly, y = Species_richness, col = SS)) + 
  geom_point() + 
  scale_color_manual(values=as.vector(alphabet())) +
  facet_wrap(~LUI) + 
  theme_bw()

ggsave(filename = paste0(outdir, "PLOT_Rich_outliers_STA_rich.png"))


## recreate plots for non-outliers

ggplot(data = pred_abun, aes(x = StdTmeanAnomaly, y = LogAbund, col = SS)) + 
  geom_point() + 
  facet_wrap(~LUI) + 
  theme_bw() + 
  theme(legend.position = "none")

ggsave(filename = paste0(outdir, "PLOT_Abun_STA_logAbun_ALL.png"))


ggplot(data = predictsSites, aes(x = StdTmeanAnomaly, y = Species_richness, col = SS)) + 
  geom_point() + 
  #scale_color_manual(values=as.vector(alphabet())) +
  facet_wrap(~LUI) + 
  theme_bw()  + 
  theme(legend.position = "none")

ggsave(filename = paste0(outdir, "PLOT_Rich_STA_rich_ALL.png"))


## focus on the outliers
ggplot(data = pred_abun, aes(x = StdTmeanAnomaly, y = LogAbund, col = SS)) + 
  geom_point() + 
  facet_wrap(~LUI) + 
  scale_color_focus(ss_abun) + 
  theme_bw() + 
  theme(legend.position = "none")

ggsave(filename = paste0(outdir, "PLOT_Abun_STA_logAbun_ALL_focus.png"))

ggplot(data = predictsSites, aes(x = StdTmeanAnomaly, y = Species_richness, col = SS)) + 
geom_point() + 
scale_color_focus(ss_rich) + 
facet_wrap(~LUI) + 
theme_bw()  + 
theme(legend.position = "none")
  
ggsave(filename = paste0(outdir, "PLOT_Rich_STA_rich_ALL_focus.png"))



## take a look at outlier studies, do some of them only have secondary vegetation?


## richness
rich_out_pred <- droplevels(filter(predictsSites, SS %in% rich_ss))

table(rich_out_pred$SS, rich_out_pred$LUI)

#                         Primary vegetation Secondary vegetation Agriculture_Low Agriculture_High
# AR1_2008__Basset 1                       0                   45               0                0
# DI1_2013__Rousseau 2                     0                   90             180               90
# HP1_2006__Lachat 1                      24                   24               0               24
# HW1_2011__Summerville 1                  0                  160               0                0
# SC1_2005__Richardson 1                   0                  295               0                0
# SC1_2011__Meijer 1                     198                   81             177                0
# SE1_2012__Poveda 1                       0                    0               0               85
# SE1_2012__Poveda 2                       0                    0               0               85
# SH1_2011__Todd 1                         0                    0               0              100
# TN1_2007__Bouyer 2                      69                   39              48               36
# AD1_2008__Billeter 6                     0                    0              64               64
# AD1_2008__Franzen 1                      0                    4              28                0
# LH1_2008__Littlewood 1                   0                   38             114                0
# SC1_2010__Marsh 1                       15                   15               0               15


# A number of these outliers only have sites in secondary vegetation

OL_sec <- c("AR1_2008__Basset 1", "HW1_2011__Summerville 1", "SC1_2005__Richardson 1")

# note, some also only have sites in high-intensity agriculture
# SE1_2012__Poveda 1, SE1_2012__Poveda 2, SH1_2011__Todd 1



# abundance

abun_out_pred <- droplevels(filter(predictsSites, SS %in% abun_ss))

table(abun_out_pred$SS, abun_out_pred$LUI)

#                         Secondary vegetation Agriculture_Low Agriculture_High
# HW1_2011__Summerville 1                  160               0                0
# SC1_2005__Richardson 1                   295               0                0
# AD1_2008__Billeter 6                       0              64               64

# the two with sec veg only are also in the rich outliers list.



## take a look across all studies, not just outliers

SS_LUI <- data.frame(table(predictsSites$SS, predictsSites$LUI))

test <- predictsSites %>% 
  group_by(SS) %>% 
  count(LUI) %>%
  pivot_wider(names_from = LUI, values_from = n)


# SC1_2005__Richardson 1 & HW1_2011__Summerville 1, both have the highlest number of sites
# in secondary vegetation across the whole dataset. 295 andn 160, the next highest is 90.

outliers <- c("SC1_2005__Richardson 1", "HW1_2011__Summerville 1")
outliers <- c("SC1_2005__Richardson 1") # 59 sites for each order
outliers <- c("HW1_2011__Summerville 1") # 80 sites for col and lep


# what is the make-up of these studies?
test_2 <- predictsSites[predictsSites$SS %in% outliers, ]

table(test_2$LUI)
# Primary vegetation Secondary vegetation      Agriculture_Low     Agriculture_High 
#                  0                  455                    0                    0 

table(test_2$Order)
# Coleoptera     Diptera   Hemiptera Hymenoptera Lepidoptera 
#        139          59          59          59         139


##%######################################################%##
#                                                          #
####      Rerun the richness models removing each       ####
####           of the outliers one at a time            ####
#                                                          #
##%######################################################%##


# for testing removing each study one at a time
outliers <- unique(predictsSites$SS) # 254 studies before outliers removed. 

# somewhere to save the effect sizes
effsizeTab <- expand.grid(
  StdTmeanAnomalyRS= c(-1.39, 0.999),
  LUI=factor(c("Primary vegetation","Secondary vegetation","Agriculture_Low","Agriculture_High"),
             levels = levels(MeanAnomalyModelRich$data$LUI)),
  Order=factor(c("Coleoptera","Diptera","Hemiptera","Hymenoptera","Lepidoptera")), 
  rem_study = outliers, 
  Eff = NA,
  lower = NA,
  upper = NA
)

# out <- outliers[1]
for(out in outliers) {
  
# subset the data to exclude the potential outliers
model_data <- droplevels(predictsSites[!predictsSites$SS == out, ]) 

MeanAnomalyModelRich <- GLMER(modelData = model_data,
                              responseVar = "Species_richness",
                              fitFamily = "poisson",
                              fixedStruct = "LUI * StdTmeanAnomalyRS * Order",
                              randomStruct = "(1|SS)+(1|SSB)+(1|SSBS)",
                              saveVars = c("SSBS"), 
                              maxIters = 20000)


# predict the effect at 1 STA for each LUI/Order

# set up table for predictions
nd <- expand.grid(
  StdTmeanAnomalyRS= c(-1.39, 0.999),
  LUI=factor(c("Primary vegetation","Secondary vegetation","Agriculture_Low","Agriculture_High"),
             levels = levels(MeanAnomalyModelRich$data$LUI)),
  Order=factor(c("Coleoptera","Diptera","Hemiptera","Hymenoptera","Lepidoptera")))

# back transform the predictors
nd$StdTmeanAnomaly <- BackTransformCentreredPredictor(
  transformedX = nd$StdTmeanAnomalyRS,
  originalX = predictsSites$StdTmeanAnomaly)

nd$Species_richness <- 0


# predict the results
pred_vals <- PredictGLMERRandIter(model = MeanAnomalyModelRich$model, data = nd)

# back transform the abundance values
pred_vals <- exp(pred_vals)

# reference for % difference = primary vegetation and positive anomaly closest to 0
refRow <- which((nd$LUI =="Primary vegetation") & (nd$StdTmeanAnomaly == min(abs(nd$StdTmeanAnomaly))))

# create list of matrices
number_of_chunks = 5
pred_vals_test <- lapply(seq(1, NROW(pred_vals), ceiling(NROW(pred_vals)/number_of_chunks)),
                              function(i) pred_vals[i:min(i + ceiling(NROW(pred_vals)/number_of_chunks) - 1, NROW(pred_vals)),])

# name them
names(pred_vals_test) <- c("Coleoptera","Diptera","Hemiptera","Hymenoptera","Lepidoptera")

# sweep out according to the refRow
pred_vals_perc <- lapply(pred_vals_test,FUN=function(x){
  sweep (x=x, MARGIN = 2, STATS=x[refRow[1],],FUN="/") 
})

list2env(pred_vals_perc,globalenv())

effs <- rbind(Coleoptera, Diptera, Hemiptera, Hymenoptera, Lepidoptera)

medvals <- apply(effs, 1, median) * 100 -100
lowers <- apply(effs, 1, quantile, 0.025) * 100 -100
uppers <- apply(effs, 1, quantile, 0.975) * 100 -100

effsizeTab[effsizeTab$rem_study == out, "Eff"] <- medvals
effsizeTab[effsizeTab$rem_study == out, "lower"] <- lowers
effsizeTab[effsizeTab$rem_study == out, "upper"] <- uppers

write.csv(effsizeTab, file = paste0(outdir, "EffsizeTab_ALL.csv"), row.names = F)

##%######################################################%##
#                                                          #
#### plot of changes in predictions as outliers removed ####
#                                                          #
##%######################################################%##

# set variables as factors
effsizeTab$rem_study <- factor(effsizeTab$rem_study, levels = outliers)
effsizeTab$Order <- as.factor(effsizeTab$Order)
effsizeTab$LUI <- factor(effsizeTab$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))
effsizeTab$STA <- NA
effsizeTab$STA[effsizeTab$StdTmeanAnomalyRS == -1.39] <- 0
effsizeTab$STA[effsizeTab$StdTmeanAnomalyRS == 0.999] <- 1
effsizeTab$STA <- as.factor(effsizeTab$STA)

p <- ggplot(effsizeTab, aes(x = LUI, y = Eff, col = rem_study)) + 
  geom_point(aes(shape = STA, group = STA), size = 1.5, position= position_dodge(width = 1)) + 
  geom_hline(yintercept = 0, linetype = "dashed", linewidth = 0.2) +
  #geom_errorbar(aes(x = LUI, ymin = lower, ymax = upper)) +
  scale_color_discrete(guide = "none") + 
  facet_wrap(~ Order) +
  xlab("") +
  ylab("Predicted percentage change") + 
  theme_bw() + 
  theme(legend.position = "bottom",
        panel.grid = element_blank(), 
        axis.text.x = element_text(angle = 65, hjust = 1))

p
ggsave(p, filename = paste0(outdir, "FIG_outlier_impact_STA10_ALL_nolabels.png"), width = 8, height = 7)

library(plotly)
ggplotly(p)


# add annotations to those with large differences

# organise the labels - outliers only plot
# dat_text <- data.frame(
#   label = c("HW1_2011__Summerville 1", "AR1_2008__Basset 1", "SC1_2005__Richardson 1", "SC1_2005__Richardson 1", "SC1_2011__Meijer 1","HW1_2011__Summerville 1", "DI1_2013__Rousseau 2"),
#   Order   = as.factor(c(rep("Coleoptera", 3), "Diptera", "Hemiptera", "Lepidoptera", "Hymenoptera")),
#   x     = c(3.4, 3.2, 3.4, 3.3, 3.2, 3.3, 3),
#   y     = c(50, 96, 500, 30, 139, 20, -25), 
#   rem_study = c("HW1_2011__Summerville 1", "AR1_2008__Basset 1", "SC1_2005__Richardson 1", "SC1_2005__Richardson 1", "SC1_2011__Meijer 1", "HW1_2011__Summerville 1", "DI1_2013__Rousseau 2")
# )

# organise the labels - all studies plot
dat_text <- data.frame(
  label = c("SC1_2005__Richardson 1", "AD1_2008__Billeter 6", "HW1_2011__Summerville 1", "SC1_2011__Meijer 1"),
  Order   = as.factor(c(rep("Hemiptera", 4))),
  x     = c(3,3,3,3),
  y     = c(210, 280, 340, 400), 
  rem_study = c("SC1_2005__Richardson 1", "AD1_2008__Billeter 6", "HW1_2011__Summerville 1", "SC1_2011__Meijer 1")
)

# add the labels to the plot
p + geom_text(
  data    = dat_text,
  mapping = aes(x = x, y = y, label = label),
 size = 3
)

# save
ggsave(filename = paste0(outdir, "FIG_outlier_impact_STA10_ALL.pdf"), width = 8, height = 7)






##%######################################################%##
#                                                          #
####      Rerun the abundance models removing each      ####
####           of the outliers one at a time            ####
#                                                          #
##%######################################################%##

# list outliers, in order of highest to lowest Cook's distance. 
outliers <- c("AD1_2008__Billeter 6", "HW1_2011__Summerville 1", "SC1_2005__Richardson 1" )
# 3 outliers for the abundance model

# somewhere to save the effect sizes
effsizeTab <- data.frame(rem_study = rep(outliers, each = 20), 
                         Order = rep(c("Coleoptera", "Diptera", "Hemiptera", "Hymenoptera", "Lepidoptera"), each = 4),
                         LUI = rep(c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High")), 
                         Eff = rep(NA),
                         lower = rep(NA),
                         upper = rep(NA))

outliers <- unique(predictsSites$SS)

# out <- outliers[1]
for(out in outliers) {
  
  # subset the data to exclude the potential outliers
  model_data <- predictsSites[!predictsSites$SS == out, ]
  
  model_data <- droplevels(model_data[!is.na(model_data$LogAbund), ])
  
  MeanAnomalyModelAbund <- GLMER(modelData = model_data, 
                                 responseVar = "LogAbund",
                                 fitFamily = "gaussian",
                                 fixedStruct = "LUI * StdTmeanAnomalyRS * Order",
                                 randomStruct = "(1|SS)+(1|SSB)",
                                 saveVars = c("SSBS"))
  
  
  
  # predict the effect at 1 STA for each LUI/Order
  
  # values for rescaled anomaly taken from the dataset
  nd <- expand.grid(
    LUI=factor(c(rep("Primary vegetation", 2),"Secondary vegetation","Agriculture_Low","Agriculture_High"),
               levels = levels(MeanAnomalyModelAbund$data$LUI)),
    Order=factor(c("Coleoptera","Diptera","Hemiptera","Hymenoptera","Lepidoptera"),
                 levels = levels(MeanAnomalyModelAbund$data$Order)))
  
  nd$StdTmeanAnomaly <- c(0.001239622, rep(1.0022959, 4))
  nd$StdTmeanAnomalyRS <- c(-1.387751, rep(0.9925570, 4))
  
  nd$LogAbund <- 0
  
  
  # predict the results
  pred_vals <- PredictGLMERRandIter(model = MeanAnomalyModelAbund$model, data = nd)
  
  # back transform the abundance values
  pred_vals <- exp(pred_vals)-0.01
  
  # reference for % difference = primary vegetation and positive anomaly closest to 0
  refRow <- which((nd$LUI=="Primary vegetation") & (nd$StdTmeanAnomaly== 0.001239622))
  
  # create list of matrices
  number_of_chunks = 5
  pred_vals_test <- lapply(seq(1, NROW(pred_vals), ceiling(NROW(pred_vals)/number_of_chunks)),
                           function(i) pred_vals[i:min(i + ceiling(NROW(pred_vals)/number_of_chunks) - 1, NROW(pred_vals)),])
  
  # name them
  names(pred_vals_test) <- c("Coleoptera","Diptera","Hemiptera","Hymenoptera","Lepidoptera")
  
  # sweep out according to the refRow
  pred_vals_perc <- lapply(pred_vals_test,FUN=function(x){
    sweep (x=x, MARGIN = 2, STATS=x[refRow[1],],FUN="/") 
  })
  
  list2env(pred_vals_perc,globalenv())
  
  effs <- rbind(Coleoptera, Diptera, Hemiptera, Hymenoptera, Lepidoptera)
  
  medvals <- apply(effs, 1, median) * 100 -100
  lowers <- apply(effs, 1, quantile, 0.025) * 100 -100
  uppers <- apply(effs, 1, quantile, 0.975) * 100 -100
  
  effsizeTab[effsizeTab$rem_study == out, "Eff"] <- medvals[c(2:5, 7:10, 12:15, 17:20, 22:25)]
  effsizeTab[effsizeTab$rem_study == out, "lower"] <- lowers[c(2:5, 7:10, 12:15, 17:20, 22:25)]
  effsizeTab[effsizeTab$rem_study == out, "upper"] <- uppers[c(2:5, 7:10, 12:15, 17:20, 22:25)]
  
  
  
}



##%######################################################%##
#                                                          #
#### plot of changes in predictions as outliers removed ####
#                                                          #
##%######################################################%##

# set variables as factors
effsizeTab$rem_study <- factor(effsizeTab$rem_study, levels = outliers)
effsizeTab$Order <- as.factor(effsizeTab$Order)
effsizeTab$LUI <- factor(effsizeTab$LUI, levels = c("Primary vegetation", "Secondary vegetation", "Agriculture_Low", "Agriculture_High"))

p <- ggplot(effsizeTab, aes(x = LUI, y = Eff, col = rem_study)) + 
  geom_point() + 
  geom_hline(yintercept = 0, linetype = "dashed", linewidth = 0.2) +
  #geom_errorbar(aes(x = LUI, ymin = lower, ymax = upper)) +
  facet_wrap(~ Order) +
  xlab("") +
  ylab("Predicted percentage change") + 
  theme_bw() + 
  theme(legend.position = "none",
        panel.grid = element_blank(), 
        axis.text.x = element_text(angle = 65, hjust = 1))

p

library(plotly)
ggplotly(p)


# add annotations to those with large differences

# organise the labels
dat_text <- data.frame(
  label = c("HW1_2011__Summerville 1", "AR1_2008__Basset 1", "SC1_2005__Richardson 1", "SC1_2005__Richardson 1", "SC1_2011__Meijer 1","HW1_2011__Summerville 1", "DI1_2013__Rousseau 2"),
  Order   = as.factor(c(rep("Coleoptera", 3), "Diptera", "Hemiptera", "Lepidoptera", "Hymenoptera")),
  x     = c(3.4, 3.2, 3.4, 3.3, 3.2, 3.3, 3),
  y     = c(50, 96, 500, 30, 139, 20, -25), 
  rem_study = c("HW1_2011__Summerville 1", "AR1_2008__Basset 1", "SC1_2005__Richardson 1", "SC1_2005__Richardson 1", "SC1_2011__Meijer 1", "HW1_2011__Summerville 1", "DI1_2013__Rousseau 2")
)

# add the labels to the plot
p + geom_text(
  data    = dat_text,
  mapping = aes(x = x, y = y, label = label),
  size = 3
)

# save
ggsave(filename = paste0(outdir, "FIG_outlier_impact_Abun.png"), width = 8, height = 7)




## Map the sites that were identified as outliers

# Set the path to copy of the database
predicts.path <- paste0("data/database.rds")

# Read in the PREDICTS data
predicts <- ReadPREDICTS(predicts.path)
# 3250404 obs. of 67 variables

# Select only data for insects
predicts <- predicts[(predicts$Class=="Insecta"),]
# 935078 obs. of 67 variables

# remove the 4 studies identified as influential outliers
outliers <- c("SC1_2005__Richardson 1", "HW1_2011__Summerville 1", "SC1_2011__Meijer 1", "AD1_2008__Billeter 6")
predicts <- predicts[predicts$SS %in% outliers, ] # 118584 obs

# Correct effort-sensitive abundance measures (assumes linear relationship between 
# effort and recorded abundance)
predicts <- CorrectSamplingEffort(diversity = predicts)
# Correcting 0 missing sampling effort values
# Rescaling sampling effort
# Correcting 734398 values for sensitivity to sampling effort

# Merge sites that have the same coordinates, from the same study and same taxonomic 
# family (e.g. multiple traps on a single transect)
predicts <- predictsFunctions::MergeSites(diversity = predicts)

# keep top five orders 
predicts <- predicts %>% filter(Order %in% c("Hymenoptera", "Coleoptera", "Lepidoptera", "Diptera", "Hemiptera")) %>% droplevels()
# 106603 obs. of 67  variables

# convert Order to a "factor"
predicts$Order <- as.factor(predicts$Order)


# use split function to split the predicts data frame into 5 data frames (1/Order)
OrderName <- paste0("",predicts$Order)

by_Order <- split(predicts,OrderName)

# extract data frames from list into global environment
list2env(by_Order,globalenv())

# drop levels and calculate site level metrics
Coleoptera <- droplevels(Coleoptera)
Coleoptera <- SiteMetrics(diversity = Coleoptera,
                          extra.cols = c("Predominant_land_use",
                                         "SSB","SSBS", "Biome","Order","UN_subregion","Best_guess_binomial"))

Diptera <- droplevels(Diptera)
Diptera <- SiteMetrics(diversity = Diptera,
                       extra.cols = c("Predominant_land_use",
                                      "SSB","SSBS", "Biome","Order","UN_subregion","Best_guess_binomial"))

Hemiptera <- droplevels(Hemiptera)
Hemiptera <- SiteMetrics(diversity = Hemiptera,
                         extra.cols = c("Predominant_land_use",
                                        "SSB","SSBS", "Biome","Order","UN_subregion","Best_guess_binomial"))

Hymenoptera <- droplevels(Hymenoptera)
Hymenoptera <- SiteMetrics(diversity = Hymenoptera,
                           extra.cols = c("Predominant_land_use",
                                          "SSB","SSBS", "Biome","Order","UN_subregion","Best_guess_binomial"))

Lepidoptera <- droplevels(Lepidoptera)
Lepidoptera <- SiteMetrics(diversity = Lepidoptera,
                           extra.cols = c("Predominant_land_use",
                                          "SSB","SSBS", "Biome","Order","UN_subregion","Best_guess_binomial"))

# merge all sites_Order data frames into one called "sites"
sites <- rbind(Coleoptera, Diptera, Hemiptera, Hymenoptera, Lepidoptera)
# 1039 rows

# First, we will rearrange the land-use classification 
# rename "Predominant_land_use" to "LandUse"
sites$LandUse <- paste(sites$Predominant_land_use)

# Drop classification where land use could not be identified
sites$LandUse[(sites$LandUse=="Cannot decide")] <- NA

# drop classification where use intensity could not be identified
sites$Use_intensity[sites$Use_intensity=="Cannot decide"] <- NA

# Now make the variable a factor, and set the reference level to primary vegetation
sites$LandUse <- factor(sites$LandUse)
sites$LandUse <- relevel(sites$LandUse,ref="Primary vegetation")

# combine LandUse (LU) and Use Intensity (UI) into new variable Land Use Intensity (LUI)
sites$LUI <- paste0(sites$LandUse,'_',sites$Use_intensity)
sites$LUI[grep("NA",sites$LUI)] <- NA # where "NA" appears in the UI field, drop classification

# recode according to land use and use intensity combinations
sites$LUI <- dplyr::recode(sites$LUI,
                           'Primary vegetation_Minimal use' = 'Primary vegetation',
                           'Cropland_Light use' = 'Agriculture_High',
                           'Secondary vegetation (indeterminate age)_Minimal use' = 'Secondary vegetation',
                           'Urban_Light use' = 'Urban',
                           'Secondary vegetation (indeterminate age)_Light use' = 'Secondary vegetation',
                           'Cropland_Intense use' = 'Agriculture_High',
                           'Cropland_Minimal use' = 'Agriculture_Low',
                           'Pasture_Light use' = 'Agriculture_Low',
                           'Pasture_Minimal use' = 'Agriculture_Low',
                           'Intermediate secondary vegetation_Minimal use' = 'Secondary vegetation',
                           'Mature secondary vegetation_Minimal use' = 'Secondary vegetation',
                           'Secondary vegetation (indeterminate age)_Intense use' = 'Secondary vegetation',
                           'Pasture_Intense use' = 'Agriculture_High',
                           'Urban_Minimal use' = 'Urban',
                           'Primary vegetation_Light use' = 'Primary vegetation',
                           'Young secondary vegetation_Light use' = 'Secondary vegetation',
                           'Urban_Intense use' = 'Urban',
                           'Primary vegetation_Intense use' = 'Primary vegetation',
                           'Young secondary vegetation_Minimal use' = 'Secondary vegetation',
                           'Mature secondary vegetation_Intense use' = 'Secondary vegetation',
                           'Plantation forest_Minimal use' = 'Agriculture_Low',
                           'Plantation forest_Intense use' = 'Agriculture_High',
                           'Young secondary vegetation_Intense use' = 'Secondary vegetation',
                           'Plantation forest_Light use' = 'Agriculture_High',
                           'Mature secondary vegetation_Light use' = 'Secondary vegetation',
                           'Intermediate secondary vegetation_Intense use' = 'Secondary vegetation',
                           'Intermediate secondary vegetation_Light use' = 'Secondary vegetation')

# 1039 obs. of 28 variables

# remove the urban sites and sites that are NA in LUI
sites <- sites[!sites$LUI == "Urban", ]
sites <- sites[!is.na(sites$LUI), ]
# 7574 obs. of 28 variables

sites <- droplevels(sites)


# plot the raster in ggplot
map.world <- map_data('world')


## 1. plot map of site distribution: scatterplot ##

# plot the raster in ggplot

# map of sites, no colour
p_map1.1 <- ggplot() +
  geom_polygon(data = map.world, aes(x = long, y = lat, group = group), alpha = 0.5,  fill = c("grey")) +
  coord_fixed() +
  geom_point(data = sites, aes(x = Longitude, y = Latitude, col = Order), shape = 21, size=1.5, col = c("#104E8B"), fill = c("#104E8B")) +
  theme(axis.title.x = element_blank(),
        axis.title.y = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank(),
        plot.background = element_blank(),
        panel.background = element_blank(),
        legend.position = "none")

ggsave(filename = "10_Additional_Tests/Outliers/Map_4_outliers.png")
